Makefile: Make config check for u_boot target
Currently one can bypass the check that a defconfig has been selected
by directly using the u_boot target. As we add more targets that
depend on u_boot those too will bypass the check. As u_boot is the
component that needs the defconfig, move the check to it.

Signed-off-by: Andrew Davis <afd@ti.com>