Makefile: Add an SD card image generation target
Creates a simple image with a 32MB FAT16 partition that can be copied to
an SD card and used to boot to U-Boot prompt or run a uEnv.txt script.

The ROM also supports FAT32 but we use FAT16 here as it is more simple
and performant, and has everything we need on a small boot partition
like this.

Signed-off-by: Andrew Davis <afd@ti.com>