Makefile: Extend sdcard target to include flashing to on-board flash
sdcard creates an image with a boot partition with the standard firmware so
the board can boot from the sdcard, as well as an ESP partition containing
the firmware in capsules and a uEnv.txt file will set EFI variables that
initiates a capsule update. Upon the next boot (using the boot partition if
booting from the sdcard), the updated firmware in the capsules will be
flashed onto the board.

So this image can be used both in the scenario where the on-board flash has
never been flashed with the firmware (or it is bricked), as well as when booting
from on-board flash and wanting to update that firmware.

Also adds a file that contains the commands to initiate the capsule update
process.  This file will be copied into the boot partition of the image as the
uEnv.txt file.

Signed-off-by: Jonathan Humphreys <j-humphreys@ti.com>