configs: Use BOARD_NAME for the board name
This used to be used for k3-image-gen and it needed to always be "evm".
Now that we are adding split defconfigs in U-Boot for EVM and SK boards
reuse this old variable to store the real board name so we can
differentiate boards using the same SoC and security type.

Start with the Beagle boards.

Signed-off-by: Andrew Davis <afd@ti.com>